R code
In theory this should successfully read in tweets, identify a search term (after the words "tweet about"), and reply with the message "My programming isn't done yet, but you asked me to tweet about [SEARCH TERM]". Currently, it makes the call to see new DMs and comes back with none no matter how many I send, which is a problem.